requirements.txt
opencv-python
numpy
imutils
matplotlib
scipy
dlib
pillow